LayoutTests/fast/speech/scripted/emma-expected.txt

Test the SpeechRecognitionResult.emma attribute.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS 'webkitSpeechRecognition' in self is true
PASS webkitSpeechRecognition == null is false
onstart
onaudiostart
onsoundstart
onspeechstart
onresult
PASS event.emma.childNodes.length is 1
PASS event.emma.firstChild.nodeName is "emma:emma"
window.emma = event.emma.firstChild
PASS emma.getAttribute("version") is "1.0"
PASS emma.childNodes.length is 1
PASS emma.firstChild.nodeName is "emma:one-of"
window.oneof = emma.firstChild
window.emmaNs = 'http://www.w3.org/2003/04/emma'
PASS oneof.getAttribute("emma:medium") is "acoustic"
PASS oneof.getAttributeNS(emmaNs, "medium") is "acoustic"
PASS oneof.getAttribute("emma:mode") is "voice"
PASS oneof.getAttributeNS(emmaNs, "mode") is "voice"
PASS oneof.childNodes.length is 1
PASS oneof.firstChild.nodeName is "emma:interpretation"
window.interpretation = oneof.firstChild
PASS interpretation.getAttribute("id") is "1"
PASS interpretation.getAttribute("emma:confidence") is "0.42"
PASS interpretation.getAttributeNS(emmaNs, "confidence") is "0.42"
PASS interpretation.getElementsByTagName("literal").length is 1
window.literal = interpretation.getElementsByTagName("literal")[0]
PASS literal.childNodes.length is 1
PASS literal.firstChild.data is "hello, world"

Test that emma's DOM tree stays alive:
window.emma = null
window.oneof = null
window.interpretation = null
window.literal = null
event.emma.customProperty = 42
gc()
PASS event.emma.customProperty is 42
onspeechend
onsoundend
onaudioend
onend
PASS successfullyParsed is true

TEST COMPLETE
